Polished the code, moved structs to the distinct files, reconstructed methods in the DataController, changed confirmation buttons
