+++ EmotionDesign.xcodeproj/xcuserdata/Ooops.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,54 +3,4 @@
    uuid = "2A250B78-2CA4-48D3-8FFE-AC849AEE2255"
    type = "1"
    version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "791BBDD1-95D5-44E6-A7A8-F2D841D8F8D6"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "EmotionDesign/Model/ColorMap.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "29"
-            endingLineNumber = "29"
-            landmarkName = "ColorMap"
-            landmarkType = "13">
-            <Locations>
-               <Location
-                  uuid = "791BBDD1-95D5-44E6-A7A8-F2D841D8F8D6 - fa0b4e32a7694528"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "EmotionDesign.ColorMap.rawValue.getter : Swift.String"
-                  moduleName = "EmotionDesign"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/Ooops/Developer/EmotionDesign/EmotionDesign/Model/ColorMap.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "27"
-                  endingLineNumber = "27"
-                  offsetFromSymbolStart = "432">
-               </Location>
-               <Location
-                  uuid = "791BBDD1-95D5-44E6-A7A8-F2D841D8F8D6 - 83d0765d066ee3ae"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "EmotionDesign.ColorMap.init(rawValue: Swift.String) -&gt; Swift.Optional&lt;EmotionDesign.ColorMap&gt;"
-                  moduleName = "EmotionDesign"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/Ooops/Developer/EmotionDesign/EmotionDesign/Model/ColorMap.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "27"
-                  endingLineNumber = "27"
-                  offsetFromSymbolStart = "533">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
 </Bucket>
